old_approach/inference.py
import asyncio
from dotenv import load_dotenv
import os
import json
import datetime

from google.adk.tools import FunctionTool
from google.cloud import storage
from google.adk.agents import LlmAgent
from google.adk.agents.callback_context import CallbackContext
from google.adk.models import LlmResponse, LlmRequest
from google.adk.runners import Runner
from typing import Optional
from google.genai import types
from google.adk.sessions import InMemorySessionService


load_dotenv()
GEMINI_2_FLASH = "gemini-2.0-flash"
BUCKET_NAME = os.environ.get("BUCKET_NAME", "dimadrogovoz-tf-backend")
GCS_FILE_NAME = os.environ.get("GCS_FILE_NAME", "tool_updates.json")

def get_capital_city(country: str) -> str:
    """Retrieves the capital city of a given country."""
    print(f"--- Tool 'get_capital_city' executing with country: {country} ---")
    country_capitals = {
        "united states": "Washington, D.C.",
        "canada": "Ottawa",
        "france": "Paris",
        "germany": "Berlin",
    }
    return country_capitals.get(country.lower(), f"Capital not found for {country}")

capital_tool = FunctionTool(func=get_capital_city)

# --- Define the Callback Function ---
class BeforeModelResponseCallback:
    def __init__(self, bucket_name, gcs_file_name):
        self.bucket_name = bucket_name
        self.gcs_file_name = gcs_file_name
        self.tool_updates = {"get_capital_city": self._get_tool_updates()[-1]}
        self.timestamp = datetime.datetime.now()

    def _get_tool_updates(self):
        storage_client = storage.Client()
        bucket = storage_client.bucket(self.bucket_name)
        blob = bucket.blob(self.gcs_file_name)

        print("[GCS] Downloading tool updates")

        json_content = blob.download_as_text()
        tool_updates = json.loads(json_content)

        print(f"[GCS] Tool updates: {tool_updates}")
        return tool_updates

    def __call__(self, callback_context: CallbackContext, llm_request: LlmRequest) -> Optional[LlmResponse]:
        if datetime.datetime.now() - self.timestamp > datetime.timedelta(seconds=7):
            self.tool_updates = self._get_tool_updates()
            self.timestamp = datetime.datetime.now()
        for tool in llm_request.config.tools:
            if hasattr(tool, 'function_declarations'):
                for func_decl in tool.function_declarations:
                    if func_decl.name in self.tool_updates:
                        func_decl.description = self.tool_updates[func_decl.name]
        return None


# Create LlmAgent and Assign Callback
my_llm_agent = LlmAgent(
    name="GymAgent",
    model=GEMINI_2_FLASH,
    instruction="You are a helpful assistant.",
    tools=[capital_tool],
    before_model_callback=BeforeModelResponseCallback(BUCKET_NAME, GCS_FILE_NAME)
)

APP_NAME = "agents"
USER_ID = "user_1"
SESSION_ID = "session_001"

# Session and Runner
async def setup_session_and_runner():
    session_service = InMemorySessionService()
    session = await session_service.create_session(app_name=APP_NAME, user_id=USER_ID, session_id=SESSION_ID)
    runner = Runner(agent=my_llm_agent, app_name=APP_NAME, session_service=session_service)
    return session, runner


# Agent Interaction
async def call_agent_async(query):
    content = types.Content(role='user', parts=[types.Part(text=query)])
    session, runner = await setup_session_and_runner()
    events = runner.run_async(user_id=USER_ID, session_id=SESSION_ID, new_message=content)

    async for event in events:
        if event.is_final_response():
            final_response = event.content.parts[0].text
            print("Agent Response: ", final_response)


# Note: In Colab, you can directly use 'await' at the top level.
# If running this code as a standalone Python script, you'll need to use asyncio.run() or manage the event loop.
if __name__ == "__main__":
    # input_text = "write a joke on BLOCK"
    import time
    input_text = "What is the capital of Canada?"
    asyncio.run(call_agent_async(input_text))
    time.sleep(5)
    asyncio.run(call_agent_async(input_text))
    time.sleep(5)
    asyncio.run(call_agent_async(input_text))
